before change to spring mvc framework

+++ src/main/java/shopping/repository/impl/CartItemDaoImpl.java
@@ -1,5 +1,7 @@
 package shopping.repository.impl;
 
+import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.List;
 
 import javax.persistence.NoResultException;
@@ -10,6 +12,7 @@ import org.hibernate.SessionFactory;
 import _00.utils.HibernateUtils;
 import createAccount.model.MemberBean;
 import shopping.model.CartItem;
+import shopping.model.GroupBuyBean;
 import shopping.model.Product;
 import shopping.repository.CartItemDao;
 
@@ -72,25 +75,37 @@ public class CartItemDaoImpl implements CartItemDao {
 	@Override
 	public void delete(Product productBean, MemberBean memberBean) {
 		Session session = factory.getCurrentSession();
-		CartItem cartItem = null;
+		List<CartItem> cartItems = null;
 		String hql = "From CartItem WHERE memberBean = :memberBean and product = :product";
-		cartItem = (CartItem) session.createQuery(hql).setParameter("memberBean", memberBean)
-				.setParameter("product", productBean).getSingleResult();
-		session.delete(cartItem);
+		cartItems = session.createQuery(hql).setParameter("memberBean", memberBean).setParameter("product", productBean)
+				.getResultList();
+
+		for (Iterator iterator = cartItems.iterator(); iterator.hasNext();) {
+			CartItem cartItem = (CartItem) iterator.next();
+			if (cartItem.getGroupBuyBean() == null) {
+				session.delete(cartItem);
+			}
+		}
 	}
 
 	@Override
 	public List<CartItem> checkAllItems(MemberBean memberBean) {
 		List<CartItem> results = null;
+		List<CartItem> cartItems = new ArrayList<CartItem>();
 		Session session = factory.getCurrentSession();
 		try {
 			String hql = "FROM CartItem WHERE memberBean = :memberBean";
 			results = session.createQuery(hql).setParameter("memberBean", memberBean).getResultList();
+			for (Iterator iterator = results.iterator(); iterator.hasNext();) {
+				CartItem cartItem = (CartItem) iterator.next();
+				if (cartItem.getGroupBuyBean() == null) {
+					cartItems.add(cartItem);
+				}
+			}
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
-		return results;
-
+		return cartItems;
 	}
 
 	@Override
@@ -125,6 +140,52 @@ public class CartItemDaoImpl implements CartItemDao {
 			System.out.println("此會員購物車沒有該筆商品");
 		}
 		return cartItemPersist;
+	}
+
+	@Override
+	public List<CartItem> checkAllItems(GroupBuyBean groupBuyBean, MemberBean memberBean) {
+		Session session = factory.getCurrentSession();
+		List<CartItem> cartItems = null;
+		try {
+			String hql = "FROM CartItem WHERE groupBuyBean = :groupBuyBean and memberBean = :memberBean";
+			cartItems = session.createQuery(hql).setParameter("groupBuyBean", groupBuyBean)
+					.setParameter("memberBean", memberBean).getResultList();
+//			if (cartItemPersist != null) {
+//				session.evict(cartItemPersist);
+//			}
+		} catch (NoResultException e) {
+			System.out.println("此會員購物車沒有該筆商品");
+		}
+		return cartItems;
+	}
+
+	@Override
+	public CartItem checkItem(Product productBean, MemberBean memberBean, GroupBuyBean groupBuyBean) {
+		Session session = factory.getCurrentSession();
+		CartItem cartItemPersist = null;
+		try {
+			String hql = "FROM CartItem WHERE product = :product and memberBean = :memberBean and groupBuyBean=:groupBuyBean";
+			cartItemPersist = (CartItem) session.createQuery(hql).setParameter("product", productBean)
+					.setParameter("memberBean", memberBean).setParameter("groupBuyBean", groupBuyBean)
+					.getSingleResult();
+//			if (cartItemPersist != null) {
+//				session.evict(cartItemPersist);
+//			}
+		} catch (NoResultException e) {
+			System.out.println("此會員購物車沒有該筆商品");
+		}
+		return cartItemPersist;
+	}
+
+	@Override
+	public void delete(Product productBean, MemberBean memberBean, GroupBuyBean groupBuyBean) {
+		Session session = factory.getCurrentSession();
+		CartItem cartItem = null;
+		String hql = "From CartItem WHERE memberBean = :memberBean and product = :product and groupBuyBean=: groupBuyBean";
+		cartItem = (CartItem) session.createQuery(hql).setParameter("memberBean", memberBean)
+				.setParameter("groupBuyBean", groupBuyBean).setParameter("product", productBean).getSingleResult();
+		session.delete(cartItem);
+
 	}
 
 }
